forming shape of whole repo
