+++ README.md
@@ -1,2 +1,19 @@
 # stimulation
 Tools for testing stimulation protocols, including millisecond timing
+
+## utilities
+
+Various helper utilities for stimulus-related code. __NB:__ make sure the `utilities`-folder is in your python path. If you have made a copy of this repository, _e.g._, into your projects' `misc`-folder, your should add the following line to code using the utilities
+
+```python
+import sys
+sys.path.insert(0, '/projects/YOUR_MINDLABPROJ/misc/stimulation/utilities')
+```
+
+## notebooks
+
+Jupyter notebooks using/demonstrating various features.
+
+## psychopy_cfin
+
+PsychoPy-scripts using/demonstrating various features.
