--- __init__.py
@@ -1,2 +1,0 @@
-from .psychopy_cfin import *
-from .utilities import *
