+++ psychopy_cfin/megadmin/find_att_cutoff_amp.py
@@ -5,15 +5,15 @@ Background
 ----------
 
 The sound card of the stimulation PC has a higher voltage output compared with
-the input voltage limit of the CFIN audio attenuator. This script allows for a manual
-routine to find the current limit. *This test only needs to be performed once, after
-any change to the sound-generating hardware.*
+the input voltage limit of the CFIN audio attenuator. This script allows for a
+manual routine to find the current limit. *This test only needs to be performed
+once, after any change to the sound-generating hardware.*
 
 Method
 ------
 
-Run the script, which sends simple 100Hz sounds (100 ms) at 2Hz. Control the sound card
-output volume using these keys:
+Run the script, which sends simple 100Hz sounds (100 ms) at 2Hz. Control the
+sound card output volume using these keys:
     u: step up the volume by the current _diff_ amount (starts at 0.1)
     d: step down the volume by the current _diff_ amount (starts at 0.1)
     1: set the current _diff_ amount to 0.1 (10**-1)
@@ -35,13 +35,13 @@ from numpy import log10
 from psychopy import prefs  # must be imported first!
 prefs.general['audioLib'] = ['pyo']
 
-from psychopy import sound, core, visual, parallel, event
+from psychopy import sound, core, visual, parallel, event  # noqa
 sound.init(rate=44100, stereo=True, buffer=128)
-print 'Using %s(with %s) for sounds' %(sound.audioLib, sound.audioDriver)
+print 'Using %s(with %s) for sounds' % (sound.audioLib, sound.audioDriver)
 
 testSound = sound.Sound(value=100, secs=0.1, stereo=True, volume=0.5, loops=0,
-                        sampleRate=44100, bits=16, hamming=False, start=0, stop=-1,
-                        name='', autoLog=False)
+                        sampleRate=44100, bits=16, hamming=False, start=0,
+                        stop=-1, name='', autoLog=False)
 
 parallel.setPortAddress(0xDFF8)  # for stimpc-08, August 2016
 try:
@@ -55,7 +55,7 @@ else:
     setParallelData = parallel.setData
 
 
-win = visual.Window(monitor='testMonitor', units ='deg', fullscr=False)
+win = visual.Window(monitor='testMonitor', units='deg', fullscr=False)
 message1 = visual.TextStim(win, pos=[0, +3], text='Press any key to start...')
 message2 = visual.TextStim(win, pos=[0, -3], text='')
 
@@ -80,7 +80,7 @@ while bContinue:
     setParallelData(0)
 
     core.wait(0.500)
-    
+
     anyKey = event.getKeys(keyList=allActiveKeys)
     if not anyKey:
         continue
@@ -88,7 +88,7 @@ while bContinue:
         bContinue = False
     elif anyKey[0] in ['u']:
         curVol = testSound.getVolume() + curDiff
-        testSound.setVolume(curVol)        
+        testSound.setVolume(curVol)
     elif anyKey[0] in ['d']:
         curVol = testSound.getVolume() - curDiff
         testSound.setVolume(curVol)
