+++ utilities/__init__.py
@@ -1,0 +1,3 @@
+from .psychopy_utils import attenuator
+from . import wavhelpers
+# cannot import extract_delays, since it depends on mne not av. in psychopy?
