+++ utilities/psychopy_utils/__init__.py
@@ -1,0 +1,1 @@
+from .attenuator import (AttenuatorController, FakeAttenuatorController)
