+++ utilities/psychopy_utils/attenuator.py
